Fixed integration tests (local and in CI)
- Added hvac service integration test
- Added external component config for vs.code tasks and containers (`prerequisite_settings.json`)
- vs.code tasks improved, added integration test (local) for dapr testing and integration test(containers) to test with released containers
- Removed `--task` argument checks for vs.code task scripts (`run-*.sh`)
- Removed versions from `requirements-dev.txt`
- Added custom volume dir for `dbc2val` configuration (original config does not work with databroker)
- Fixed metadata bug in `vdb_helper` causing `ValueErrors` with `VEHICLEDATABROKER_DAPR_APP_ID` set
- Fixed docker container testing (`integration_test/it-*` scripts)
- `it-setup.sh start` made lazy by default (if containers are started, do not re-start them)
- Added some timeout after container startup, needed for feedercan datapoint registration
- Enabled local containers and binaries rebuild if using special "prerelease" version for seat and hvac servces. In that case `docker-build.sh --local x86_64` is used for building and tagging images. Also works on CI with additional tools installed.
- Fixed dependabot versions in github actions

+++ integration_test/it-setup.sh
@@ -19,8 +19,6 @@
 
 SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 
-set -e
-
 # shellcheck source=/dev/null
 source ${SCRIPT_DIR}/it-config
 
@@ -29,20 +27,24 @@ cleanup() {
 	local force="$1"
 	# ensure containers are not running
 	echo "# Cleanup iteration test containers..."
-	if docker ps -a | grep "${VDB_CONTAINER}"; then
-		docker container rm -f "${VDB_CONTAINER}"
+	if docker ps -a | grep "${KDB_CONTAINER}"; then
+		docker container rm -f "${KDB_CONTAINER}"
 	fi
 	if docker ps -a | grep "${SEAT_CONTAINER}"; then
 		docker container rm -f "${SEAT_CONTAINER}"
+	fi
+	if docker ps -a | grep "${HVAC_CONTAINER}"; then
+		docker container rm -f "${HVAC_CONTAINER}"
 	fi
 	if docker ps -a | grep "${FEEDER_CONTAINER}"; then
 		docker container rm -f "${FEEDER_CONTAINER}"
 	fi
 	if [ "$force" = "1" ]; then
 		echo "# Cleanup VAL ghcr images..."
-		docker image rm -f "${VDB_IMAGE}"
-		docker image rm -f "${SEAT_IMAGE}"
 		docker image rm -f "${FEEDER_IMAGE}"
+		docker image rm -f "${KDB_IMAGE}"
+		[ "$SEAT_TAG" != "prerelease" ] && docker image rm -f "${SEAT_IMAGE}"
+		[ "$HVAC_TAG" != "prerelease" ] && docker image rm -f "${HVAC_IMAGE}"
 	fi
 }
 
@@ -62,39 +64,49 @@ pull_images() {
 		return 0
 	fi
 	if [ "${force}" = "1" ] ||
-		! __check_docker_image "${VDB_IMAGE}" ||
+		! __check_docker_image "${KDB_IMAGE}" ||
 		! __check_docker_image "${SEAT_IMAGE}" ||
+		! __check_docker_image "${HVAC_IMAGE}" ||
 		! __check_docker_image "${FEEDER_IMAGE}"; then
 		echo "- Pulling images form ${DOCKER_REPO} (May need manual login)..."
 		docker login "${DOCKER_REPO}"
 
-		docker pull "${VDB_IMAGE}"
-		docker pull "${SEAT_IMAGE}"
+		echo "- docker pull ${KDB_IMAGE}"
+		docker pull "${KDB_IMAGE}"
+		echo "- docker pull ${FEEDER_IMAGE}"
 		docker pull "${FEEDER_IMAGE}"
+		if [ "$SEAT_TAG" != "prerelease" ]; then
+			echo "- docker pull ${SEAT_IMAGE}"
+			docker pull "${SEAT_IMAGE}"
+		fi
+		if [ "$HVAC_TAG" != "prerelease" ]; then
+			echo "- docker pull ${HVAC_IMAGE}"
+			docker pull "${HVAC_IMAGE}"
+		fi
 	fi
 }
 
-build_images() {
-	local force="$1"
-	if [ "$force" = "1" ] || ! __check_docker_image "${SEAT_IMAGE}"; then
-		echo "# Building amd64/seat-service:latest ..."
-		if cd ${SCRIPT_DIR}/../seat_service && ./docker-build.sh -l x86_64; then
-			docker tag amd64/seat-service:latest ${SEAT_IMAGE}
-		fi
-	fi
-	if [ "$force" = "1" ] || ! __check_docker_image "${VDB_IMAGE}"; then
-		echo "# Building amd64/databroker:latest ..."
-		if cd ${SCRIPT_DIR}/../kuksa_data_broker && ./docker-build.sh -l x86_64; then
-			docker tag amd64/databroker:latest ${VDB_IMAGE}
-		fi
-	fi
-	if [ "$force" = "1" ] || ! __check_docker_image "${FEEDER_IMAGE}"; then
-		echo "# Building amd64/feeder_can:latest ..."
-		if cd ${SCRIPT_DIR}/../feeder_can && ./docker-build.sh -l x86_64; then
-			docker tag amd64/feeder_can:latest ${FEEDER_IMAGE}
-		fi
-	fi
-}
+# build_images() {
+# 	local force="$1"
+# 	if [ "$force" = "1" ] || ! __check_docker_image "${SEAT_IMAGE}"; then
+# 		echo "# Building amd64/seat-service:latest ..."
+# 		if cd ${SCRIPT_DIR}/../seat_service && ./docker-build.sh -l x86_64; then
+# 			docker tag amd64/seat-service:latest ${SEAT_IMAGE}
+# 		fi
+# 	fi
+# 	if [ "$force" = "1" ] || ! __check_docker_image "${KDB_IMAGE}"; then
+# 		echo "# Building amd64/databroker:latest ..."
+# 		if cd ${SCRIPT_DIR}/../vehicle_data_broker && ./docker-build.sh -l x86_64; then
+# 			docker tag amd64/databroker:latest ${KDB_IMAGE}
+# 		fi
+# 	fi
+# 	if [ "$force" = "1" ] || ! __check_docker_image "${FEEDER_IMAGE}"; then
+# 		echo "# Building amd64/feeder_can:latest ..."
+# 		if cd ${SCRIPT_DIR}/../feeder_can && ./docker-build.sh -l x86_64; then
+# 			docker tag amd64/feeder_can:latest ${FEEDER_IMAGE}
+# 		fi
+# 	fi
+# }
 
 ### Checks if container (name) is running
 __check_container_state() {
@@ -120,42 +132,50 @@ __check_docker_image() {
 	fi
 }
 
-### Checks if $VDB_CONTAINER and $SEAT_CONTAINER are both running
+### Checks if $KDB_CONTAINER and $SEAT_CONTAINER are both running
 check_it_containers() {
 	local verbose="$1"
 
 	local seat_err=0
-	local vdb_err=0
+	local kdb_err=0
 	local feed_err=0
-	__check_container_state "${VDB_CONTAINER}" "${verbose}" || vdb_err=1
+	local hvac_err=0
+
+	__check_container_state "${KDB_CONTAINER}" "${verbose}" || kdb_err=1
 	__check_container_state "${SEAT_CONTAINER}" "${verbose}" || seat_err=1
+	__check_container_state "${HVAC_CONTAINER}" "${verbose}" || hvac_err=1
 	__check_container_state "${FEEDER_CONTAINER}" "${verbose}" || feed_err=1
 
-	if [ ${vdb_err} -ne 0 ] || [ ${seat_err} -ne 0 ] || [ ${feed_err} -ne 0 ]; then
+	if [ ${kdb_err} -ne 0 ] || [ ${seat_err} -ne 0 ] || [ ${feed_err} -ne 0 ] || [ ${hvac_err} -ne 0 ]; then
 		return 1
 	fi
 	return 0
 }
 
-### Starts $VDB_CONTAINER and $SEAT_CONTAINER
+### Starts $KDB_CONTAINER and $SEAT_CONTAINER
 start_containers() {
-	echo "- Running ${VDB_CONTAINER} ..."
+	echo "- Running ${KDB_CONTAINER} ..."
 	# DataBroker container options
 	rc=0
-	docker run -d ${DOCKER_OPT} ${VDB_DOCKER_OPT} "${VDB_IMAGE}" || rc=1
+	docker run -d ${DOCKER_OPT} ${KDB_DOCKER_OPT} "${KDB_IMAGE}" || rc=1
 
 	echo "- Running ${SEAT_CONTAINER} ..."
 	# SeatService container options. BROKER_ADDR is needed to reach it-databroker ports within val-test network
 	docker run -d ${DOCKER_OPT} ${SEAT_DOCKER_OPT} "${SEAT_IMAGE}" || rc=2
 
+	echo "- Running ${HVAC_CONTAINER} ..."
+	# SeatService container options. BROKER_ADDR is needed to reach it-databroker ports within val-test network
+	docker run -d ${DOCKER_OPT} ${HVAC_DOCKER_OPT} "${HVAC_IMAGE}" || rc=3
+
 	echo "- Running ${FEEDER_CONTAINER} ..."
 	# SeatService container options. BROKER_ADDR is needed to reach it-databroker ports within val-test network
-	docker run -d ${DOCKER_OPT} ${FEEDER_DOCKER_OPT} "${FEEDER_IMAGE}" || rc=3
+	docker run -d ${DOCKER_OPT} ${FEEDER_DOCKER_OPT} "${FEEDER_IMAGE}" || rc=4
 
 	echo
-	__check_container_state "${VDB_CONTAINER}" 1 || rc=1
+	__check_container_state "${KDB_CONTAINER}" 1 || rc=1
 	__check_container_state "${SEAT_CONTAINER}" 1 || rc=2
-	__check_container_state "${FEEDER_CONTAINER}" 1 || rc=3
+	__check_container_state "${HVAC_CONTAINER}" 1 || rc=3
+	__check_container_state "${FEEDER_CONTAINER}" 1 || rc=4
 	echo
 
 	return ${rc}
@@ -189,7 +209,7 @@ it_init() {
 	fi
 
 	# auto pull/build images (only if missing)
-	if echo "${VDB_TAG}${SEAT_TAG}${FEEDER_TAG}" | grep -q "latest"; then
+	if echo "${KDB_TAG}${SEAT_TAG}${HVAC_TAG}${FEEDER_TAG}" | grep -q "latest"; then
 		build_images "${force}"
 	else
 		pull_images "${force}"
@@ -201,20 +221,23 @@ it_init() {
 
 it_start() {
 	local force="$1"
-	set -e
 
-	# initial cleanup for images/containers
-	cleanup "$force"
+	if [ "$force" = "1" ]; then
+		# initial cleanup for images/containers
+		cleanup "$force"
+	fi
 
 	# ensure images are pulled, containers are started
 	it_init "$force"
 
-	start_containers
+	if [ "$force" = "1" ] || ! check_it_containers; then
+		start_containers
+	fi
+
 	return $?
 }
 
 it_stop() {
-	set -e
 
 	# cleanup for containres only
 	cleanup 0
@@ -232,12 +255,17 @@ it_status() {
 	echo "-----------------------"
 	echo
 	if [ "${logs}" = "1" ]; then
-		echo "### [${VDB_CONTAINER}] Logs:"
-		docker logs $DOCKER_LOG "${VDB_CONTAINER}"
+		echo "### [${KDB_CONTAINER}] Logs:"
+		docker logs $DOCKER_LOG "${KDB_CONTAINER}"
 		echo "-----------------------"
 		echo
 		echo "### [${SEAT_CONTAINER}] Logs:"
 		docker logs $DOCKER_LOG "${SEAT_CONTAINER}"
+		echo "-----------------------"
+		echo
+		echo
+		echo "### [${HVAC_CONTAINER}] Logs:"
+		docker logs $DOCKER_LOG "${HVAC_CONTAINER}"
 		echo "-----------------------"
 		echo
 		echo
@@ -271,7 +299,7 @@ it_usage() {
 	echo "  cleanup  Removes VAL Containers. Use --force to also remove configured VAL images"
 }
 
-# FIXME: BROKER_ADDR=${VDB_CONTAINER}:55555 should be changed in case of host network...
+# FIXME: BROKER_ADDR=${KDB_CONTAINER}:55555 should be changed in case of host network...
 export DOCKER_OPT="--network ${DOCKER_NETWORK} ${DOCKER_OPT}"
 
 # parse options in $FORCE and $CMD
